Немного пофиксил расчёт оценки.

+++ Elgamal/WindowsFormsApplication1/TutorResults.cs
@@ -48,7 +48,7 @@ namespace ElgamalTutor
             else
                 ReverseAnsLabel2.Text += "Неверно";
             //Some LINQ magic
-            int summa = Answers.modpowAnswers.Sum(a => a ? 1 : 0) + Answers.EulerAnswers.Sum(a => a ? 1 : 0) + Answers.ReverseAnswers.Sum(a => a ? 1 : 0);
+            double summa = Answers.modpowAnswers.Sum(a => a ? 1 : 0) + Answers.EulerAnswers.Sum(a => a ? 1 : 0) + Answers.ReverseAnswers.Sum(a => a ? 1 : 0);
             double FinalMark = 3 * (summa / 8) + 2;
             if (FinalMark % 1 >= 0.5)
                 FinalMark += 1;
